Rename ChatDropdownWeb to ChatDropdown and update imports
The component and file ChatDropdownWeb were renamed to ChatDropdown for clarity and consistency. The context is that originally this just housed web functionality, but as more features were added it was repurposed into a general store for features beyond file upload.

All relevant imports and usage references were updated to reflect the new name.
